Small fixes

+++ fractol.h
@@ -71,7 +71,7 @@ void	ft_primary_arg_check(int argc, char **argv);
 
 /*Draw utilities*/
 void	put_color_to_pixel(t_mlx *mlx, int x, int y, int color);
-int		draw_fractal(t_mlx *mlx, char *choice, int x, int y);
+int		draw_fractal(t_mlx *mlx, char *choice, double x, double y);
 int		pick_color(int i, t_mlx *mlx);
 
 /*Mandelbrot*/
@@ -79,7 +79,7 @@ void	*draw_mandl(void *mlx_void);
 void    calc_mandl(t_mlx *mlx);
 
 /*Julia*/
-void    draw_julia(t_mlx *mlx, int x, int y);
-void    calc_julia(t_mlx *mlx, int x, int y);
+void    draw_julia(t_mlx *mlx, double x, double y);
+void    calc_julia(t_mlx *mlx, double x, double y);
 
 #endif
